# go.mod: upgrade Prometheus