--- a/engine/sort.go
+++ b/engine/sort.go
@@ -19,6 +19,10 @@ const (
 	sortOrderDesc sortOrder = true
 )
 
+type keepHistogramsSorter interface {
+	keepHistograms()
+}
+
 type resultSorter interface {
 	comparer(samples *promql.Vector) func(i, j int) bool
 }
@@ -42,6 +46,10 @@ type aggregateResultSort struct {
 
 type noSortResultSort struct {
 }
+
+func (a aggregateResultSort) keepHistograms() {}
+
+func (s noSortResultSort) keepHistograms() {}
 
 func extractSortingLabels(f *parser.Call) []string {
 	args := f.Args[1:]
@@ -102,6 +110,17 @@ func valueCompare(order sortOrder, l, r float64) bool {
 		return l < r
 	}
 	return l > r
+}
+
+// filterFloats filters out histogram samples from the vector in-place.
+func filterFloats(v promql.Vector) promql.Vector {
+	floats := v[:0]
+	for _, s := range v {
+		if s.H == nil {
+			floats = append(floats, s)
+		}
+	}
+	return floats
 }
 
 func (s sortFuncResultSort) comparer(samples *promql.Vector) func(i, j int) bool {