--- a/execution/aggregate/count_values.go
+++ b/execution/aggregate/count_values.go
@@ -11,11 +11,13 @@ import (
 	"sync"
 	"time"
 
+	"github.com/efficientgo/core/errors"
+	prommodel "github.com/prometheus/common/model"
+	"github.com/prometheus/prometheus/model/labels"
+
 	"github.com/thanos-io/promql-engine/execution/model"
 	"github.com/thanos-io/promql-engine/execution/telemetry"
 	"github.com/thanos-io/promql-engine/query"
-
-	"github.com/prometheus/prometheus/model/labels"
 )
 
 type countValuesOperator struct {
@@ -116,6 +118,10 @@ func (c *countValuesOperator) Next(ctx context.Context) ([]model.StepVector, err
 }
 
 func (c *countValuesOperator) initSeriesOnce(ctx context.Context) error {
+	if !prommodel.LabelName(c.param).IsValid() {
+		return errors.Newf("invalid label name %q", c.param)
+	}
+
 	nextSeries, err := c.next.Series(ctx)
 	if err != nil {
 		return err