Upgrade the Prometheus library and the corresponding code.