@GiedriusS 2025-05-22
Somehow `rate` and double exponential smoothing are much worse after the update:
```
goos: linux
goarch: amd64
pkg: github.com/thanos-io/promql-engine/engine
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz
                                                      │ benchmarks/new.out │      benchmarks_update/new.out       │
                                                      │       sec/op       │    sec/op      vs base               │
RangeQuery/vector_selector-16                                 68.53m ± ∞ ¹    74.67m ± ∞ ¹   +8.96% (p=0.008 n=5)
RangeQuery/sum-16                                             41.39m ± ∞ ¹    49.03m ± ∞ ¹  +18.46% (p=0.008 n=5)
RangeQuery/sum_by_pod-16                                      68.52m ± ∞ ¹    91.31m ± ∞ ¹  +33.27% (p=0.008 n=5)
RangeQuery/topk-16                                            42.56m ± ∞ ¹    52.44m ± ∞ ¹  +23.21% (p=0.008 n=5)
RangeQuery/bottomk-16                                         43.22m ± ∞ ¹    50.66m ± ∞ ¹  +17.21% (p=0.008 n=5)
RangeQuery/limitk-16                                          42.11m ± ∞ ¹    50.75m ± ∞ ¹  +20.51% (p=0.008 n=5)
RangeQuery/limit_ratio-16                                     175.8m ± ∞ ¹    294.7m ± ∞ ¹  +67.65% (p=0.008 n=5)
RangeQuery/rate-16                                            90.97m ± ∞ ¹   106.88m ± ∞ ¹  +17.49% (p=0.008 n=5)
RangeQuery/rate_with_longer_window-16                         45.98m ± ∞ ¹    56.47m ± ∞ ¹  +22.82% (p=0.008 n=5)
RangeQuery/subquery-16                                        189.5m ± ∞ ¹    283.1m ± ∞ ¹  +49.41% (p=0.008 n=5)
RangeQuery/sum_rate-16                                        77.53m ± ∞ ¹    87.20m ± ∞ ¹  +12.48% (p=0.008 n=5)
RangeQuery/sum_by_rate-16                                     85.87m ± ∞ ¹   100.75m ± ∞ ¹  +17.33% (p=0.008 n=5)
RangeQuery/quantile_with_variable_parameter-16                94.46m ± ∞ ¹   117.17m ± ∞ ¹  +24.04% (p=0.008 n=5)
RangeQuery/binary_operation_with_one_to_one-16                61.77m ± ∞ ¹    79.22m ± ∞ ¹  +28.24% (p=0.008 n=5)
RangeQuery/binary_operation_with_many_to_one-16               154.7m ± ∞ ¹    177.8m ± ∞ ¹  +14.91% (p=0.008 n=5)
RangeQuery/binary_operation_with_vector_and_scalar-16         104.6m ± ∞ ¹    112.2m ± ∞ ¹   +7.25% (p=0.008 n=5)
RangeQuery/unary_negation-16                                  70.86m ± ∞ ¹    73.36m ± ∞ ¹   +3.54% (p=0.008 n=5)
RangeQuery/vector_and_scalar_comparison-16                    105.7m ± ∞ ¹    124.9m ± ∞ ¹  +18.21% (p=0.008 n=5)
RangeQuery/positive_offset_vector-16                          66.05m ± ∞ ¹    81.59m ± ∞ ¹  +23.54% (p=0.008 n=5)
RangeQuery/at_modifier_-16                                    43.10m ± ∞ ¹    54.73m ± ∞ ¹  +26.97% (p=0.008 n=5)
RangeQuery/at_modifier_with_positive_offset_vector-16         43.02m ± ∞ ¹    55.52m ± ∞ ¹  +29.05% (p=0.008 n=5)
RangeQuery/clamp-16                                           89.27m ± ∞ ¹   102.21m ± ∞ ¹  +14.50% (p=0.008 n=5)
RangeQuery/clamp_min-16                                       82.99m ± ∞ ¹    98.46m ± ∞ ¹  +18.64% (p=0.008 n=5)
RangeQuery/complex_func_query-16                              122.6m ± ∞ ¹    162.7m ± ∞ ¹  +32.70% (p=0.008 n=5)
RangeQuery/func_within_func_query-16                          62.47m ± ∞ ¹    73.69m ± ∞ ¹  +17.96% (p=0.008 n=5)
RangeQuery/aggr_within_func_query-16                          54.65m ± ∞ ¹    69.37m ± ∞ ¹  +26.95% (p=0.008 n=5)
RangeQuery/histogram_quantile-16                              394.0m ± ∞ ¹    624.0m ± ∞ ¹  +58.38% (p=0.008 n=5)
RangeQuery/sort-16                                            69.19m ± ∞ ¹    89.43m ± ∞ ¹  +29.27% (p=0.008 n=5)
RangeQuery/sort_desc-16                                       69.95m ± ∞ ¹    82.41m ± ∞ ¹  +17.82% (p=0.008 n=5)
RangeQuery/absent_and_exists-16                               39.66m ± ∞ ¹    49.98m ± ∞ ¹  +26.00% (p=0.008 n=5)
RangeQuery/absent_and_doesnt_exist-16                         1.499m ± ∞ ¹    2.330m ± ∞ ¹  +55.44% (p=0.008 n=5)
RangeQuery/double_exponential_smoothing-16                    83.41m ± ∞ ¹   116.91m ± ∞ ¹  +40.16% (p=0.008 n=5)
NativeHistograms/selector-16                                  404.4m ± ∞ ¹    518.4m ± ∞ ¹  +28.18% (p=0.008 n=5)
NativeHistograms/sum-16                                       520.8m ± ∞ ¹    665.5m ± ∞ ¹  +27.78% (p=0.008 n=5)
NativeHistograms/rate-16                                      398.9m ± ∞ ¹    550.0m ± ∞ ¹  +37.86% (p=0.008 n=5)
NativeHistograms/rate_with_longer_window-16                   210.5m ± ∞ ¹    283.9m ± ∞ ¹  +34.88% (p=0.008 n=5)
NativeHistograms/sum_rate-16                                  465.6m ± ∞ ¹    542.8m ± ∞ ¹  +16.58% (p=0.008 n=5)
NativeHistograms/histogram_sum-16                             846.6m ± ∞ ¹   1029.5m ± ∞ ¹  +21.60% (p=0.008 n=5)
NativeHistograms/histogram_count_with_rate-16                 761.5m ± ∞ ¹    907.2m ± ∞ ¹  +19.13% (p=0.008 n=5)
NativeHistograms/histogram_count-16                           866.3m ± ∞ ¹    990.4m ± ∞ ¹  +14.32% (p=0.008 n=5)
NativeHistograms/histogram_count_with_sum_and_rate-16         341.0m ± ∞ ¹    372.0m ± ∞ ¹   +9.08% (p=0.008 n=5)
NativeHistograms/histogram_avg-16                             864.4m ± ∞ ¹   1012.4m ± ∞ ¹  +17.12% (p=0.008 n=5)
NativeHistograms/histogram_avg_with_sum_and_rate-16           340.3m ± ∞ ¹    390.0m ± ∞ ¹  +14.60% (p=0.008 n=5)
NativeHistograms/histogram_quantile-16                        552.1m ± ∞ ¹    602.6m ± ∞ ¹   +9.15% (p=0.008 n=5)
NativeHistograms/histogram_scalar_binop-16                    999.4m ± ∞ ¹   1088.3m ± ∞ ¹   +8.90% (p=0.008 n=5)
NativeHistograms/histogram_stdvar-16                           2.222 ± ∞ ¹     2.387 ± ∞ ¹        ~ (p=0.095 n=5)
NativeHistograms/histogram_stddev-16                           2.228 ± ∞ ¹     2.352 ± ∞ ¹   +5.57% (p=0.008 n=5)
geomean                                                       138.5m          169.8m        +22.59%
¹ need >= 6 samples for confidence interval at level 0.95

                                                      │ benchmarks/new.out │       benchmarks_update/new.out       │
                                                      │        B/op        │     B/op       vs base                │
RangeQuery/vector_selector-16                                25.28Mi ± ∞ ¹   25.47Mi ± ∞ ¹         ~ (p=0.095 n=5)
RangeQuery/sum-16                                            6.465Mi ± ∞ ¹   6.494Mi ± ∞ ¹         ~ (p=0.151 n=5)
RangeQuery/sum_by_pod-16                                     13.92Mi ± ∞ ¹   13.97Mi ± ∞ ¹    +0.38% (p=0.008 n=5)
RangeQuery/topk-16                                           8.269Mi ± ∞ ¹   8.490Mi ± ∞ ¹    +2.67% (p=0.008 n=5)
RangeQuery/bottomk-16                                        8.219Mi ± ∞ ¹   8.491Mi ± ∞ ¹    +3.30% (p=0.008 n=5)
RangeQuery/limitk-16                                         8.219Mi ± ∞ ¹   8.405Mi ± ∞ ¹         ~ (p=0.056 n=5)
RangeQuery/limit_ratio-16                                    16.24Mi ± ∞ ¹   16.31Mi ± ∞ ¹         ~ (p=0.310 n=5)
RangeQuery/rate-16                                           26.90Mi ± ∞ ¹   28.29Mi ± ∞ ¹    +5.17% (p=0.008 n=5)
RangeQuery/rate_with_longer_window-16                        14.64Mi ± ∞ ¹   14.64Mi ± ∞ ¹         ~ (p=0.310 n=5)
RangeQuery/subquery-16                                       31.07Mi ± ∞ ¹   31.05Mi ± ∞ ¹         ~ (p=0.421 n=5)
RangeQuery/sum_rate-16                                       10.16Mi ± ∞ ¹   10.01Mi ± ∞ ¹    -1.51% (p=0.008 n=5)
RangeQuery/sum_by_rate-16                                    17.06Mi ± ∞ ¹   17.06Mi ± ∞ ¹         ~ (p=1.000 n=5)
RangeQuery/quantile_with_variable_parameter-16               19.97Mi ± ∞ ¹   20.14Mi ± ∞ ¹    +0.88% (p=0.008 n=5)
RangeQuery/binary_operation_with_one_to_one-16               14.21Mi ± ∞ ¹   14.22Mi ± ∞ ¹         ~ (p=0.421 n=5)
RangeQuery/binary_operation_with_many_to_one-16              34.85Mi ± ∞ ¹   34.84Mi ± ∞ ¹         ~ (p=0.841 n=5)
RangeQuery/binary_operation_with_vector_and_scalar-16        30.10Mi ± ∞ ¹   30.16Mi ± ∞ ¹         ~ (p=0.310 n=5)
RangeQuery/unary_negation-16                                 27.90Mi ± ∞ ¹   27.97Mi ± ∞ ¹         ~ (p=0.056 n=5)
RangeQuery/vector_and_scalar_comparison-16                   29.55Mi ± ∞ ¹   29.92Mi ± ∞ ¹         ~ (p=0.056 n=5)
RangeQuery/positive_offset_vector-16                         25.93Mi ± ∞ ¹   25.98Mi ± ∞ ¹    +0.19% (p=0.016 n=5)
RangeQuery/at_modifier_-16                                   22.38Mi ± ∞ ¹   22.38Mi ± ∞ ¹         ~ (p=0.056 n=5)
RangeQuery/at_modifier_with_positive_offset_vector-16        22.38Mi ± ∞ ¹   22.38Mi ± ∞ ¹         ~ (p=0.841 n=5)
RangeQuery/clamp-16                                          27.34Mi ± ∞ ¹   27.33Mi ± ∞ ¹         ~ (p=1.000 n=5)
RangeQuery/clamp_min-16                                      27.31Mi ± ∞ ¹   27.36Mi ± ∞ ¹         ~ (p=0.056 n=5)
RangeQuery/complex_func_query-16                             30.89Mi ± ∞ ¹   31.06Mi ± ∞ ¹    +0.56% (p=0.016 n=5)
RangeQuery/func_within_func_query-16                         8.185Mi ± ∞ ¹   8.178Mi ± ∞ ¹         ~ (p=0.056 n=5)
RangeQuery/aggr_within_func_query-16                         7.932Mi ± ∞ ¹   7.925Mi ± ∞ ¹         ~ (p=0.222 n=5)
RangeQuery/histogram_quantile-16                             48.72Mi ± ∞ ¹   48.76Mi ± ∞ ¹         ~ (p=0.548 n=5)
RangeQuery/sort-16                                           26.66Mi ± ∞ ¹   26.78Mi ± ∞ ¹    +0.45% (p=0.008 n=5)
RangeQuery/sort_desc-16                                      26.62Mi ± ∞ ¹   26.77Mi ± ∞ ¹    +0.58% (p=0.008 n=5)
RangeQuery/absent_and_exists-16                              8.213Mi ± ∞ ¹   8.272Mi ± ∞ ¹         ~ (p=0.222 n=5)
RangeQuery/absent_and_doesnt_exist-16                        424.0Ki ± ∞ ¹   424.4Ki ± ∞ ¹    +0.09% (p=0.016 n=5)
RangeQuery/double_exponential_smoothing-16                   28.21Mi ± ∞ ¹   61.12Mi ± ∞ ¹  +116.64% (p=0.008 n=5)
NativeHistograms/selector-16                                 491.7Mi ± ∞ ¹   491.6Mi ± ∞ ¹         ~ (p=0.310 n=5)
NativeHistograms/sum-16                                      474.2Mi ± ∞ ¹   474.2Mi ± ∞ ¹         ~ (p=0.690 n=5)
NativeHistograms/rate-16                                     217.0Mi ± ∞ ¹   216.6Mi ± ∞ ¹    -0.17% (p=0.008 n=5)
NativeHistograms/rate_with_longer_window-16                  44.99Mi ± ∞ ¹   58.01Mi ± ∞ ¹         ~ (p=0.151 n=5)
NativeHistograms/sum_rate-16                                 213.1Mi ± ∞ ¹   212.4Mi ± ∞ ¹    -0.31% (p=0.008 n=5)
NativeHistograms/histogram_sum-16                            379.0Mi ± ∞ ¹   378.9Mi ± ∞ ¹         ~ (p=0.151 n=5)
NativeHistograms/histogram_count_with_rate-16                163.1Mi ± ∞ ¹   163.0Mi ± ∞ ¹         ~ (p=0.151 n=5)
NativeHistograms/histogram_count-16                          378.9Mi ± ∞ ¹   378.9Mi ± ∞ ¹         ~ (p=0.548 n=5)
NativeHistograms/histogram_count_with_sum_and_rate-16        144.0Mi ± ∞ ¹   143.9Mi ± ∞ ¹         ~ (p=0.841 n=5)
NativeHistograms/histogram_avg-16                            379.0Mi ± ∞ ¹   378.8Mi ± ∞ ¹         ~ (p=0.421 n=5)
NativeHistograms/histogram_avg_with_sum_and_rate-16          144.0Mi ± ∞ ¹   143.9Mi ± ∞ ¹         ~ (p=0.151 n=5)
NativeHistograms/histogram_quantile-16                       488.6Mi ± ∞ ¹   488.6Mi ± ∞ ¹         ~ (p=0.095 n=5)
NativeHistograms/histogram_scalar_binop-16                   669.8Mi ± ∞ ¹   669.5Mi ± ∞ ¹         ~ (p=0.222 n=5)
NativeHistograms/histogram_stdvar-16                         750.1Mi ± ∞ ¹   750.2Mi ± ∞ ¹         ~ (p=0.222 n=5)
NativeHistograms/histogram_stddev-16                         750.0Mi ± ∞ ¹   750.2Mi ± ∞ ¹         ~ (p=1.000 n=5)
geomean                                                      41.52Mi         42.60Mi          +2.61%
¹ need >= 6 samples for confidence interval at level 0.95

                                                      │ benchmarks/new.out │       benchmarks_update/new.out       │
                                                      │     allocs/op      │   allocs/op    vs base                │
RangeQuery/vector_selector-16                                 50.61k ± ∞ ¹    50.87k ± ∞ ¹         ~ (p=0.095 n=5)
RangeQuery/sum-16                                             48.04k ± ∞ ¹    48.08k ± ∞ ¹         ~ (p=0.095 n=5)
RangeQuery/sum_by_pod-16                                      67.46k ± ∞ ¹    67.53k ± ∞ ¹    +0.12% (p=0.008 n=5)
RangeQuery/topk-16                                            45.98k ± ∞ ¹    46.21k ± ∞ ¹    +0.50% (p=0.008 n=5)
RangeQuery/bottomk-16                                         45.94k ± ∞ ¹    46.21k ± ∞ ¹    +0.59% (p=0.008 n=5)
RangeQuery/limitk-16                                          45.71k ± ∞ ¹    45.87k ± ∞ ¹         ~ (p=0.056 n=5)
RangeQuery/limit_ratio-16                                     182.8k ± ∞ ¹    182.8k ± ∞ ¹         ~ (p=0.841 n=5)
RangeQuery/rate-16                                            74.72k ± ∞ ¹    91.69k ± ∞ ¹   +22.70% (p=0.008 n=5)
RangeQuery/rate_with_longer_window-16                         84.63k ± ∞ ¹    84.52k ± ∞ ¹         ~ (p=0.151 n=5)
RangeQuery/subquery-16                                        111.7k ± ∞ ¹    111.6k ± ∞ ¹    -0.09% (p=0.008 n=5)
RangeQuery/sum_rate-16                                        86.26k ± ∞ ¹    85.99k ± ∞ ¹    -0.32% (p=0.008 n=5)
RangeQuery/sum_by_rate-16                                     105.1k ± ∞ ¹    105.0k ± ∞ ¹         ~ (p=0.095 n=5)
RangeQuery/quantile_with_variable_parameter-16                147.9k ± ∞ ¹    147.9k ± ∞ ¹         ~ (p=0.151 n=5)
RangeQuery/binary_operation_with_one_to_one-16                56.65k ± ∞ ¹    56.61k ± ∞ ¹         ~ (p=0.690 n=5)
RangeQuery/binary_operation_with_many_to_one-16               102.2k ± ∞ ¹    102.0k ± ∞ ¹         ~ (p=0.222 n=5)
RangeQuery/binary_operation_with_vector_and_scalar-16         77.98k ± ∞ ¹    77.92k ± ∞ ¹         ~ (p=0.095 n=5)
RangeQuery/unary_negation-16                                  79.68k ± ∞ ¹    79.67k ± ∞ ¹         ~ (p=0.310 n=5)
RangeQuery/vector_and_scalar_comparison-16                    68.93k ± ∞ ¹    69.07k ± ∞ ¹         ~ (p=0.421 n=5)
RangeQuery/positive_offset_vector-16                          59.02k ± ∞ ¹    59.02k ± ∞ ¹         ~ (p=0.571 n=5)
RangeQuery/at_modifier_-16                                    48.47k ± ∞ ¹    48.44k ± ∞ ¹    -0.07% (p=0.008 n=5)
RangeQuery/at_modifier_with_positive_offset_vector-16         48.48k ± ∞ ¹    48.44k ± ∞ ¹    -0.07% (p=0.008 n=5)
RangeQuery/clamp-16                                           77.57k ± ∞ ¹    77.48k ± ∞ ¹         ~ (p=0.151 n=5)
RangeQuery/clamp_min-16                                       77.16k ± ∞ ¹    77.13k ± ∞ ¹         ~ (p=0.690 n=5)
RangeQuery/complex_func_query-16                              88.00k ± ∞ ¹    88.02k ± ∞ ¹         ~ (p=0.151 n=5)
RangeQuery/func_within_func_query-16                          87.55k ± ∞ ¹    87.44k ± ∞ ¹    -0.13% (p=0.008 n=5)
RangeQuery/aggr_within_func_query-16                          87.55k ± ∞ ¹    87.42k ± ∞ ¹    -0.15% (p=0.008 n=5)
RangeQuery/histogram_quantile-16                              531.9k ± ∞ ¹    532.0k ± ∞ ¹    +0.01% (p=0.016 n=5)
RangeQuery/sort-16                                            67.68k ± ∞ ¹    67.76k ± ∞ ¹    +0.12% (p=0.032 n=5)
RangeQuery/sort_desc-16                                       67.64k ± ∞ ¹    67.75k ± ∞ ¹    +0.17% (p=0.016 n=5)
RangeQuery/absent_and_exists-16                               61.71k ± ∞ ¹    61.67k ± ∞ ¹         ~ (p=0.310 n=5)
RangeQuery/absent_and_doesnt_exist-16                         3.725k ± ∞ ¹    3.728k ± ∞ ¹    +0.08% (p=0.008 n=5)
RangeQuery/double_exponential_smoothing-16                    85.86k ± ∞ ¹   805.73k ± ∞ ¹  +838.41% (p=0.008 n=5)
NativeHistograms/selector-16                                  8.050M ± ∞ ¹    8.050M ± ∞ ¹         ~ (p=0.222 n=5)
NativeHistograms/sum-16                                       8.046M ± ∞ ¹    8.046M ± ∞ ¹         ~ (p=0.889 n=5)
NativeHistograms/rate-16                                      3.930M ± ∞ ¹    3.903M ± ∞ ¹    -0.68% (p=0.008 n=5)
NativeHistograms/rate_with_longer_window-16                   682.6k ± ∞ ¹    689.2k ± ∞ ¹         ~ (p=0.151 n=5)
NativeHistograms/sum_rate-16                                  3.960M ± ∞ ¹    3.933M ± ∞ ¹    -0.68% (p=0.008 n=5)
NativeHistograms/histogram_sum-16                             2.417M ± ∞ ¹    2.417M ± ∞ ¹         ~ (p=0.222 n=5)
NativeHistograms/histogram_count_with_rate-16                 1.021M ± ∞ ¹    1.021M ± ∞ ¹         ~ (p=0.421 n=5)
NativeHistograms/histogram_count-16                           2.416M ± ∞ ¹    2.417M ± ∞ ¹         ~ (p=0.548 n=5)
NativeHistograms/histogram_count_with_sum_and_rate-16         1.007M ± ∞ ¹    1.007M ± ∞ ¹         ~ (p=0.841 n=5)
NativeHistograms/histogram_avg-16                             2.417M ± ∞ ¹    2.416M ± ∞ ¹         ~ (p=0.841 n=5)
NativeHistograms/histogram_avg_with_sum_and_rate-16           1.007M ± ∞ ¹    1.007M ± ∞ ¹         ~ (p=0.548 n=5)
NativeHistograms/histogram_quantile-16                        8.081M ± ∞ ¹    8.081M ± ∞ ¹    +0.00% (p=0.032 n=5)
NativeHistograms/histogram_scalar_binop-16                    11.70M ± ∞ ¹    11.70M ± ∞ ¹         ~ (p=0.095 n=5)
NativeHistograms/histogram_stdvar-16                          8.816M ± ∞ ¹    8.816M ± ∞ ¹         ~ (p=0.286 n=5)
NativeHistograms/histogram_stddev-16                          8.816M ± ∞ ¹    8.816M ± ∞ ¹         ~ (p=0.841 n=5)
geomean                                                       243.1k          256.1k          +5.36%
¹ need >= 6 samples for confidence interval at level 0.95
```

Need to investigate & fix.